# docs/swagger.yaml
definitions:
  github_com_orders_api_api_errs.ErrorResponse:
    description: 'Модель описывает возвращаемую ошибку: код и краткое сообщение'
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  github_com_orders_api_internal_models.Delivery:
    description: Модель описывает информацию о доставщике
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      region:
        type: string
      zip:
        type: string
    required:
    - city
    - name
    - phone
    type: object
  github_com_orders_api_internal_models.Item:
    description: Модель описывает информацию о товаре в заказе
    properties:
      brand:
        type: string
      chrt_id:
        type: integer
      name:
        type: string
      nm_id:
        minimum: 0
        type: integer
      price:
        minimum: 0
        type: integer
      rid:
        type: string
      sale:
        maximum: 100
        minimum: 0
        type: integer
      size:
        type: string
      status:
        type: integer
      total_price:
        minimum: 0
        type: integer
      track_number:
        type: string
    required:
    - chrt_id
    - name
    - rid
    - track_number
    type: object
  github_com_orders_api_internal_models.Order:
    properties:
      customer_id:
        type: string
      date_created:
        type: string
      delivery:
        $ref: '#/definitions/github_com_orders_api_internal_models.Delivery'
      delivery_service:
        type: string
      entry:
        type: string
      internal_signature:
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_orders_api_internal_models.Item'
        type: array
      locale:
        type: string
      oof_shard:
        type: string
      order_uid:
        type: string
      payment:
        $ref: '#/definitions/github_com_orders_api_internal_models.Payment'
      shardkey:
        type: string
      sm_id:
        minimum: 0
        type: integer
      track_number:
        type: string
    required:
    - customer_id
    - date_created
    - delivery
    - delivery_service
    - entry
    - items
    - locale
    - order_uid
    - payment
    - shardkey
    - track_number
    type: object
  github_com_orders_api_internal_models.Payment:
    description: Модель описывает информацию о платеже
    properties:
      amount:
        minimum: 0
        type: integer
      bank:
        type: string
      currency:
        type: string
      custom_fee:
        minimum: 0
        type: integer
      delivery_cost:
        minimum: 0
        type: integer
      goods_total:
        minimum: 0
        type: integer
      payment_dt:
        type: integer
      provider:
        type: string
      request_id:
        type: string
      transaction:
        type: string
    required:
    - bank
    - currency
    - provider
    - transaction
    type: object
info:
  contact: {}
  title: WB_order API
  version: "1.0"
paths:
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Регистрирует нового пользователя
      parameters:
      - description: Order UUID
        format: uuid
        in: path
        name: order_uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_orders_api_internal_models.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_orders_api_api_errs.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_orders_api_api_errs.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_orders_api_api_errs.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - orders
swagger: "2.0"